Pulled common directory for samples; updated sample branding
