--- sample/src/main/res/layout/activity_xml_layout.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
->
-    <Button
-            android:id="@+id/progress_btn"
-            android:text="Start progress"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-    />
-
-    <ProgressBar
-            android:id="@+id/progress"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
-
-</LinearLayout>
