+++ samples/android-showcase/.gitignore
