+++ samples/android-showcase/build.gradle
