+++ samples/android-showcase/proguard-rules.pro
