+++ samples/android-showcase/src/main/AndroidManifest.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    package="dev.inkremental.sample">
+
+    <application
+        android:allowBackup="false"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:supportsRtl="true"
+        android:theme="@style/AppTheme"
+        tools:ignore="GoogleAppIndexingWarning">
+        <activity
+            android:name="dev.inkremental.sample.MainActivity"
+            android:label="@string/app_name">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity android:name="dev.inkremental.sample.FragmentActivity" android:label="Fragment sample" />
+        <activity android:name="dev.inkremental.sample.ListActivity" android:label="List sample" />
+        <activity android:name="dev.inkremental.sample.BasicsActivity" android:label="Basics sample" />
+        <activity android:name="dev.inkremental.sample.ConstraintActivity" android:label="Constraint sample" />
+        <activity android:name="dev.inkremental.sample.YogaActivity" android:label="YogaLayout sample" />
+        <activity android:name="dev.inkremental.sample.XmlActivity" android:label="Xml sample" />
+    </application>
+
+</manifest>
