+++ samples/android-showcase/src/main/kotlin/dev/inkremental/sample/FragmentActivity.kt
