+++ samples/android-showcase/src/main/res/drawable/children_background.xml
