+++ samples/android-showcase/src/main/res/layout/container.xml
