+++ samples/android-showcase/src/main/res/layout/preview.xml
