+++ samples/android-showcase/src/main/res/values/colors.xml
