+++ samples/android-showcase/src/main/res/values/styles.xml
